backend/requirements.txt
asgiref==3.8.1
Django==4.2.16
django-allauth==65.2.0
dj-rest-auth==7.0.0
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
dotenv==1.0.1
psycopg2-binary==2.9.10
requests==2.32.3
sqlparse==0.5.1
typing-extensions==4.12.2
